refactor: move exceptions to global and remove application env
