+++ app/src/main/resources/application.properties
@@ -4,6 +4,3 @@ management.endpoints.web.exposure.include=*
 # CONFIG DATABASE
 spring.mongodb.embedded.enabled=false
 spring.data.mongodb.uri=mongodb://mongodb:27017/wishlist
-
-## APP ENV
-wishlist.max.product.limit=20
