+++ app/src/test/resources/application-test.properties
@@ -1,6 +1,3 @@
 spring.data.mongodb.uri=mongodb://localhost:27017/testdb
 de.flapdoodle.mongodb.embedded.version=4.0.2
 spring.data.mongodb.database=testdb
-
-## APP ENV
-wishlist.max.product.limit=20
